windows: fix unsupported %zu

+++ examples/chatpp.cc
@@ -14,12 +14,12 @@ public:
 
 	void receive(meshlink::node *source, const void *data, size_t len) {
 		const char *msg = (const char *)data;
-	
+
 		if(!len || msg[len - 1]) {
 			fprintf(stderr, "Received invalid data from %s\n", source->name);
 			return;
 		}
-		
+
 		printf("%s says: %s\n", source->name, msg);
 	}
 
@@ -85,7 +85,7 @@ static void parse_command(meshlink::mesh *mesh, char *buf) {
 			if(!nodes) {
 				fprintf(stderr, "Could not get list of nodes: %s\n", meshlink::strerror());
 			} else {
-				printf("%zu known nodes:", nnodes);
+				printf("%lu known nodes:", (unsigned long)nnodes);
 				for(size_t i = 0; i < nnodes; i++)
 					printf(" %s", nodes[i]->name);
 				printf("\n");
